+++ src/spine/construct/fragment.py
@@ -72,6 +72,7 @@ class FragmentBuilder(BuilderBase):
         fragment_group_pred=None,
         fragment_node_pred=None,
         sources=None,
+        orig_index=None,
     ):
         """Builds :class:`RecoFragment` objects from the full chain output.
 
@@ -93,6 +94,9 @@ class FragmentBuilder(BuilderBase):
             (P) Interaction group each fragment belongs to
         sources : np.ndarray, optional
             (N, 2) Tensor which contains the module/tpc information
+        orig_index : np.ndarray, optional
+            (N) Tensor which contains the indexes in the original
+            point cloud (before any filtering or deghosting)
 
         Returns
         -------
@@ -119,6 +123,8 @@ class FragmentBuilder(BuilderBase):
             # Add optional arguments
             if sources is not None:
                 fragment.sources = sources[index]
+            if orig_index is not None:
+                fragment.orig_index = orig_index[index]
             if fragment_start_points is not None:
                 fragment.start_point = fragment_start_points[i]
             if fragment_end_points is not None and fragment.shape == TRACK_SHP:
@@ -163,6 +169,7 @@ class FragmentBuilder(BuilderBase):
         depositions_g4=None,
         sources_label=None,
         sources=None,
+        orig_index_label=None,
         particles=None,
     ):
         """Builds :class:`TruthFragment` objects from the full chain output.
@@ -196,6 +203,9 @@ class FragmentBuilder(BuilderBase):
             (N', 2) Tensor which contains the label module/tpc information
         sources : np.ndarray, optional
             (N, 2) Tensor which contains the module/tpc information
+        orig_index_label : np.ndarray, optional
+            (N') Tensor which contains the indexes in the original
+            point cloud (before any filtering or deghosting)
         particles : List[Particle], optional
             List of true particles
 
@@ -261,6 +271,8 @@ class FragmentBuilder(BuilderBase):
                     fragment.depositions_q = depositions_q_label[index_ref]
                 if sources_label is not None:
                     fragment.sources = sources_label[index_ref]
+                if orig_index_label is not None:
+                    fragment.orig_index = orig_index_label[index_ref]
 
                 # If the fragments are not broken, can match to G4 info
                 if not broken: